[ownership] Add support for handling address_to_pointer to LoadBorrowInvalidationChecker.
We treat address_to_pointer as an explicit opt in escape from the protection we
provide. The code emitter must instead provide other manners of guaranteeing
safety such as using mark_dependence. So we just treat the instruction as a
non-write instruction.

This verifier is not enabled in 5.3, so this does not need to be cherry-picked
to there.

<rdar://problem/63387309>