runtime: silence -Wunused-variable warning (NFC)
`classMetadata` is only used in the ObjC path, resulting in a
`-Wunused-variable` warning.  Sink the variable into the ObjC path.
Because the conversion of the metadata does not rely on the type
metadata bits in the metadata, it is safe to delay the definition and
bit adjustment to the point where it is used unifying the two ObjC
paths.