Introduce type annotations to permit use of one-way closure parameters.
There are two places in our normal build that are missing type
annotations that would become required with one-way closure parameters.
Introduce type annotations so that we can build enough of the code
base to perform additional testing.

These changes are unnecessary (but harmless) if nothing comes of the
one-way closure experiment.