[build-script] Explicitly separate build-script-impl products and non-build-script-impl-products when building the list of products to build.
This was already being done in a really hacky way namely, we computed the
build-script-impl list by filtering a single product list. But if one looked at
the actual product classes that we have, all of the build-script-impl products
are already a prefix of the entire original list?! So just repersent it as a
prefix/suffix pair.

I also put in an assert to verify we do not break this invariant in the future.