Add experimental support for multi-statement closure inference.
Introduce support for type inference of closures with multiple
statements. Type inference flows from the closure parameters through
the body of the closure to determine the return type. This, for example,
allows a "map" closure with multiple statements to infer the result type
based on the "return" in the body:

    let _ = ints.map { i in
      print(i)
      return String(i)
    }

The frontend option `-experimental-multi-statement-closures` enables
this new behavior. Only simple expression statements in closures work
thus far.