Warn when using Clang declarations from implicitly textually included headers.
The immediate motivation for this comes from this proposal to change the
`SwiftGlibc` module map, which will result in glibc headers being
textually included:

https://forums.swift.org/t/problems-with-swiftglibc-and-proposed-fix/37594

In this context, we want to prevent user modules from starting to export
declarations from libc headers. However, the new diagnostic is useful
generally.

Currently, the diagnostic we output is always a warning. A followup PR
will change this to an error for declarations that came from a textually
included libc header.