test: exit properly with `swift_build_sdk_interface`
`sys.exit` takes an integral value in the range of `0-127`.  Python 3 is
much more stringent about this and will convert the value if it is out
of the range to 1.  On Windows, explicitly use `ExitProcess` to ensure
that we get the correct value on the exit code which matters to `not`.