Forward matching of trailing closure arguments.
Introduce an experimental attribute and command-line flag to replace
the "backward" matching algorithm for unlabeled trailing closure
arguments with a simpler "forward" algorithm. In this new algorithm,
unlabeled trailing closure arguments match the next parameter in the
parameter list that can accept an unlabeled trailing closure.

The "can accept an unlabeled trailing closure" criteria looks at the
parameter itself. The parameter accepts an unlabeled trailing closure
if all of the following are true:

* The parameter is not variadic
* The parameter is not 'inout'
* The adjusted type of the parameter (defined below) is a function type

The adjusted type of the parameter is the parameter's type as
declared, after performing two adjustments:

* If the parameter is an @autoclosure, use the result type of the
parameter's declared (function) type, before performing the second
adjustment.
* Remove all outer "optional" types.

For example, the following function illustrates both adjustments to
determine that the parameter "body" accepts an unlabeled trailing
closure:

    func doSomething(body: @autoclosure () -> (((Int) -> String)?))

This source-breaking change is accessible in one of two ways. The new attribute

    @_trailingClosureMatching(.forward)

can be applied to a function, initializer, or subscript to enable the
new forward matching algorithm for direct calls to that function. This
allows specific APIs to opt in to forward matching while leaving all
other calls unaffected.

The command-line parameter `-enable-experimental-trailing-closure-matching`
enables forward matching as the default algorithm. It is most useful
for experimenting with the source-compatibility effects of forward
matching.

Specific APIs can also explicitly request to use the current, backward
matching algorithm to maintain their existing behavior even under the
new command-line parameter, by using the other form of the new
trailing closure matching attribute:

    @_trailingClosureMatching(.backward)