[AST] Stop using PointerIntPair for SIL params/results
The differentiability bit pushed the size of SILParameterInfo and
SILResultInfo from 8 bytes to 16 bytes. Given that the low bits of the
PointerIntPair are full and given that we do not want to over-align
types, the PointerIntPair is not saving space anymore. Let's simplify
these two data structures to make debugging easier (and the speed of
debug builds slightly faster). In theory, release builds will also
benefit due to fewer masking operations to extract the pointer or the
low bits, but this will probably be "in the noise" for benchmarks.

Note: this change is techinically not NFC because it fixes the
SILResultInfo equality comparison function to check differentiability.