build: force UTF-8 IO encoding
This should make the build system more robust in the case that `LOCALE`
is unset.  In such an environment, python will default to ASCII rather
than UTF-8 as the default locale is C, which is ASCII.

Thanks to @texasmichelle for identifying the issue and @codafi for
pointing me to `PYTHONIOENCODING`.