[5.3] Improved codgen for FP inits from Integer (#32636)
* Provide an implementation of init?<T:BinaryInteger>(exactly:T) on each stdlib FP type. (#32632)

Previously these always went through the FloatingPoint-provided default implementation, which is not particularly efficient. Also try removing inlinable from the generic _convert hooks, since we probably never want to actually inline them.

* Add explicit init from BinaryInteger to CGFloat

The override was previously missing, which meant that we fell back on the generic implementation; we should simply forward this to the NativeType implementation instead.