[Apple Silicon] [build-script] Build compiler-rt even when cross-compiling
clang and compiler-rt are rev-locked and should match.  Currently, we build clang (and use it to bootstrap the rest) but don't build compiler-rt.  We should.

This doesn't completely achieve that: this only makes us build compiler-rt for the local host.

This used to be pretty much impossible, but, since then, compiler-rt cmake support was rewritten.

We should build it for the other hosts as well, but that's trickier, because we'd need to make the cross-compiled compiler-rt build use a separate just-built clang (from the local host). We do that for the other projects, but compiler-rt is special (because it's cmake'd as a runtime part of clang/llvm).